# Conflicts:
#	application/config/config.php
#	application/modules/clients/views/view.php
#	application/modules/dashboard/views/index.php
#	application/modules/invoices/controllers/Ajax.php
#	application/modules/invoices/controllers/Cron.php
#	application/modules/invoices/views/modal_create_invoice.php
#	application/modules/invoices/views/view.php
#	application/modules/quotes/controllers/Ajax.php
#	composer.json
#	composer.lock

+++ application/modules/import/controllers/Import.php
@@ -99,6 +99,8 @@ class Import extends Admin_Controller
                             break;
                     }
                 }
+
+                $this->mdl_import->cleanup_import_files();
             }
 
             redirect('import');
@@ -110,6 +112,10 @@ class Import extends Admin_Controller
      */
     public function delete($id)
     {
+        if ( ! $this->ensure_valid_post_request('import')) {
+            return;
+        }
+
         $this->mdl_import->delete($id);
         redirect('import');
     }
